% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_partial_effect.R
\name{average_partial_effect}
\alias{average_partial_effect}
\title{Estimate average partial effect via augmented balancing}
\usage{
average_partial_effect(X, Y, W, balance.method = c("minimax", "plugin"),
  zeta = 0.5, fit.method = c("elnet", "none"), alpha = 1,
  scale.X = TRUE, solver = c("ECOS", "SCS"), verbose = TRUE)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{W}{the effect variable (real valued)}

\item{balance.method}{how the balancing weights gamma are derived}

\item{zeta}{tuning parameter for selecting approximately balancing weights}

\item{fit.method}{the method used to fit mu(x, w) = E[Y | X = x, W = w]}

\item{alpha}{tuning paramter for glmnet}

\item{scale.X}{whether non-binary features should be noramlized}

\item{solver}{convex optimzer used by CVXR for minimax weights}

\item{verbose}{whether the optimizer should print progress information}
}
\value{
ATE estimate, along with standard error estimate
}
\description{
Estimate average partial effect via augmented balancing
}
