% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimax.R
\name{balance_minimax}
\alias{balance_minimax}
\title{Estimate minimax balancing weights}
\usage{
balance_minimax(X, W, zeta, solver = c("ECOS", "SCS"), verbose = TRUE)
}
\arguments{
\item{X}{the input features}

\item{W}{the effect variable (real valued)}

\item{zeta}{tuning parameter for selecting approximately balancing weights}

\item{solver}{convex optimzer used by CVXR}

\item{verbose}{whether convex optimizer should print verbose output}
}
\value{
balancing weights
}
\description{
Estimate minimax balancing weights
}
