% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{balance_plugin}
\alias{balance_plugin}
\title{Estimate plugin balancing weights}
\usage{
balance_plugin(X, W, w.hat = NULL, alpha = 1)
}
\arguments{
\item{X}{the input features}

\item{W}{the effect variable (real valued)}

\item{w.hat}{optional precomupted estimate of E[W | X]}

\item{alpha}{tuning paramter for glmnet}
}
\value{
balancing weights
}
\description{
Estimate plugin balancing weights
}
