% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{rlasso}
\alias{rlasso}
\title{Estimate treatment effect via the R-lasso, as proposed by Nie and Wager (2017)}
\usage{
rlasso(X, Y, W, alpha = 1, nfolds = NULL, lambda.choice = c("lambda.1se",
  "lambda.min"), standardize = TRUE)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{W}{the effect variable (real valued)}

\item{alpha}{tuning parameter for the elastic net}

\item{nfolds}{number of folds for cross-fitting}

\item{lambda.choice}{how to cross-validated}

\item{standardize}{whether X should be rescaled before running the lasso}
}
\description{
Estimate treatment effect via the R-lasso, as proposed by Nie and Wager (2017)
}
