% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_partial_effect.R
\name{average_partial_effect}
\alias{average_partial_effect}
\title{Estimate average partial effect via augmented balancing}
\usage{
average_partial_effect(X, Y, W, balance.method = c("minimax", "plugin"),
  zeta = 0.5, fitted.model = NULL, alpha = 1, standardize = TRUE,
  solver = c("ECOS", "SCS"), verbose = TRUE)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{W}{the effect variable (real valued)}

\item{balance.method}{how the balancing weights gamma are derived}

\item{zeta}{tuning parameter for selecting approximately balancing weights}

\item{fitted.model}{optional pre-fitted regression adjustment}

\item{alpha}{tuning paramter for glmnet}

\item{standardize}{whether non-binary features should be noramlized}

\item{solver}{convex optimzer used by CVXR for minimax weights}

\item{verbose}{whether the optimizer should print progress information}
}
\value{
ATE estimate with standard error estimate. Also returns ``linear''
        point estimate of the form sum gamma_i Yi, as in Donoho (1994), for comparison.
}
\description{
Estimate average partial effect via augmented balancing
}
